feat: Add flags claim to JWT tokens
Add a "flags" claim to the JWT tokens. This flag currently contains bitflags for
the active permissions of the token. If a tokens is missing the flags
claim, it is defaulted to empty (no flags).

The authorization has been split up into two permission flags: ADMIN and
EDIT_SCORES. The ADMIN flag is required for all operations interacting
with the REST API. The EDIT_SCORES flag allows to edit scores of an
ongoing bracket. If a request contains a valid token but does not
satisfy the required flags the server returns a 403 Forbidden status.

By default /v2/auth/login (and /v3/auth/login) return a token with all
flags.